+++ src/editor/editor-form.ts
@@ -173,6 +173,9 @@ class Form extends BaseEditor {
         } else if (child.tagName == "HA-FORM-MULTI_SELECT") {
           child.setAttribute("style", "margin-bottom: var(--ha-form-style-multiselect-margin-bottom, 24px);");
 
+        } else if (child.tagName == "HA-FORM-SELECT") {
+          child.setAttribute("style", "margin-bottom: var(--ha-form-style-multiselect-margin-bottom, 24px); --mdc-menu-item-height: 2.5rem;");
+
         } else if (child.tagName == "HA-FORM-INTEGER") {
           child.setAttribute("style", "margin-bottom: var(--ha-form-style-integer-margin-bottom, 24px);");