+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.20 ] - 2025/04/30
+
+  * Added `sectionDefault` config option to display the specified section of the card by default.  Check out the [Configuration Options wiki document](https://github.com/thlucas1/homeassistantcomponent_soundtouchplus_card/wiki/Configuration-Options#sectiondefault) for more information.
+  * Removed `footerIconSize` from UI Editor screen.  The option is still supported, just not via the UI Editor; value must be edited via the Code Editor.  Check out the [Theming Options wiki document](https://github.com/thlucas1/homeassistantcomponent_soundtouchplus_card/wiki/Themes#card-configuration-options) for more information.
+
 ###### [ 1.0.19 ] - 2025/04/28
 
   * Added sources button to minimized player so that a source can be selected prior to power on / resuming play.